# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AllYourBase API
  description: |
    Backend-as-a-Service for PostgreSQL. Single binary. Auto-generated REST API
    with filtering, sorting, pagination, foreign key expansion, full-text search,
    batch operations, authentication, file storage, realtime events, and webhooks.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: AllYourBase
    url: https://allyourbase.io
servers:
  - url: http://localhost:8090
    description: Local development server

# Default: no auth required. Individual operations override with BearerAuth or AdminAuth.
security:
  - {}
...
paths:
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks
      operationId: listWebhooks
      security:
        - AdminAuth: []
      responses:
        "200":
          description: All webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Webhooks]
      summary: Create a webhook
      operationId: createWebhook
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          description: Invalid request (missing URL or invalid event)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/webhooks/{id}:
    get:
      tags: [Webhooks]
      summary: Get a webhook
      operationId: getWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "200":
          description: The webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Webhooks]
      summary: Update a webhook (partial)
      operationId: updateWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
      responses:
        "200":
          description: The updated webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          description: Invalid event type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      operationId: deleteWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "204":
          description: Webhook deleted
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/webhooks/{id}/test:
    post:
      tags: [Webhooks]
      summary: Test webhook delivery
      description: |
        Sends a test event to the webhook URL and reports the result.
        Does not respect the enabled flag — you can test disabled webhooks.
        Single attempt with 10-second timeout, no retries.
      operationId: testWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "200":
          description: Test result (success or failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTestResult"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/login or /api/auth/register
    AdminAuth:
      type: http
      scheme: bearer
      description: Admin token from /api/admin/auth
  parameters:
...
    WebhookId:
      name: id
      in: path
      required: true
      description: Webhook UUID
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "collection not found: nonexistent"
        data:
          type: object
          description: Field-level validation details
          additionalProperties: true
...
    WebhookRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: URL to receive webhook events
        secret:
          type: string
          description: HMAC-SHA256 signing secret for payload verification
        events:
          type: array
          items:
            type: string
            enum: [create, update, delete]
          description: "Event types to subscribe to (default: all)"
          default: [create, update, delete]
        tables:
          type: array
          items:
            type: string
          description: "Table names to filter (default: all tables)"
          default: []
        enabled:
          type: boolean
          description: Whether the webhook is active
          default: true
    WebhookResponse:
      type: object
      required: [id, url, hasSecret, events, tables, enabled, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        hasSecret:
          type: boolean
          description: Whether a signing secret is configured (secret value is never exposed)
        events:
          type: array
          items:
            type: string
            enum: [create, update, delete]
        tables:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WebhookTestResult:
      type: object
      required: [success, durationMs]
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
          description: HTTP status code from target (omitted on connection error)
        durationMs:
          type: integer
          format: int64
        error:
          type: string
          description: Error message (omitted on success)
...